Part_C/app/requirements.txt
streamlit
pandas
pillow
plotly
Pyrebase4==4.7.0
urllib3==1.26.15 
requests-toolbelt==0.10.1